requirements.txt
Pillow==9.2.0
Flask~=2.1.3
tensorflow==1.14.0
elasticsearch==7.7.0
numpy==1.16.5
keras==2.3.1